Move getSVG into a separate file to further refactoring. All internal dependencies (getrgb, getType) are reimported.
